services/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg>=0.29.0
alembic==1.13.2
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic[email]
python-jose[cryptography]==3.3.0
PyJWT[crypto]==2.9.0
httpx==0.27.2
anthropic==0.37.0
pgvector==0.3.4
python-multipart==0.0.9
websockets==12.0
redis==5.1.1
pypdf==4.3.1
python-docx==1.1.2
python-pptx==1.0.2
openpyxl==3.1.5
Pillow>=10.0.0
pytest==8.3.3
pytest-asyncio==0.24.0
